Moved old maths methods to new project
